+++ src/lightwin/core/commands/chopper.py
@@ -3,33 +3,32 @@
 import logging
 
 from lightwin.core.commands.command import Command
-from lightwin.core.elements.element import Element
 from lightwin.core.instruction import Instruction
 from lightwin.tracewin_utils.line import DatLine
 
 
-class Steerer(Command):
+class Chopper(Command):
     """Dummy class."""
 
     is_implemented = False
+    n_attributes = 6
 
     def __init__(
         self, line: DatLine, dat_idx: int | None = None, **kwargs: str
     ) -> None:
         """Call the mother ``__init__`` method."""
         super().__init__(line, dat_idx)
+        self.n_elements = int(line.splitted[1])
+        self.u_v = float(line.splitted[2])
+        self.d_mm = float(line.splitted[3])
+        self.c_mm = float(line.splitted[4])
+        self.p_axis = int(line.splitted[5])
 
     def set_influenced_elements(
         self, instructions: list[Instruction], **kwargs: float
     ) -> None:
         """Determine the index of the elements concerned by :func:`apply`."""
-        next_element = list(
-            filter(
-                lambda elt: isinstance(elt, Element),
-                instructions[self.idx["dat_idx"] :],
-            )
-        )[0]
-        start = next_element.idx["dat_idx"]
+        start = self.idx["dat_idx"]
         stop = start + 1
         self.influenced = slice(start, stop)
 
@@ -37,5 +36,5 @@ class Steerer(Command):
         self, instructions: list[Instruction], **kwargs: float
     ) -> list[Instruction]:
         """Do nothing."""
-        logging.error("Steerer not implemented.")
+        logging.error("Shift not implemented.")
         return instructions