+++ src/lightwin/core/commands/superpose_map.py
@@ -11,6 +11,7 @@ from lightwin.core.elements.element import Element
 from lightwin.core.elements.field_maps.field_map import FieldMap
 from lightwin.core.elements.superposed_field_map import SuperposedFieldMap
 from lightwin.core.instruction import Instruction
+from lightwin.tracewin_utils.line import DatLine
 
 
 class SuperposeMap(Command):
@@ -26,10 +27,12 @@ class SuperposeMap(Command):
     is_implemented = True
     n_attributes = (1, 6)
 
-    def __init__(self, line: list[str], dat_idx: int, **kwargs: str) -> None:
+    def __init__(
+        self, line: DatLine, dat_idx: int | None = None, **kwargs: str
+    ) -> None:
         """Save position as attribute."""
         super().__init__(line, dat_idx)
-        self.z_0 = float(line[1]) * 1e-3
+        self.z_0 = float(line.splitted[1]) * 1e-3
 
     def set_influenced_elements(
         self, instructions: list[Instruction], **kwargs: float
@@ -133,7 +136,7 @@ class SuperposeMap(Command):
             if superposed_field_map_is_already_inserted:
                 instructions_to_merge[i] = DummyElement(*args)
                 instructions_to_merge[i].nature = "SUPERPOSED_FIELD_MAP"
-                instructions_to_merge[i].new_rf_field = RfField()
+                instructions_to_merge[i].rf_field = RfField()
                 number_of_superposed += 1
                 continue
 
@@ -141,7 +144,7 @@ class SuperposeMap(Command):
                 *args, total_length=total_length
             )
             instructions_to_merge[i].nature = "SUPERPOSED_FIELD_MAP"
-            instructions_to_merge[i].new_rf_field = RfField()
+            instructions_to_merge[i].rf_field = RfField()
             superposed_field_map_is_already_inserted = True
         return instructions_to_merge, number_of_superposed