+++ CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `evaluator` objects are more robust and can be configured from the `.toml`.
 - Plotting is now performed thanks to the `plotter` library.
 
+## [0.?.??] 2024-??-?? -- implement_field_map_70 branch
+
+### Added
+- `FIELD_MAP 70` will not raise error (warning issued with `Envelope3D`).
+
+### Changed
+- Creation of `DatLine` object, holding a line of the ``.dat`` file. Solves several bugs, e.g. with hyphens in personalized names.
+
 ## [0.8.0b0] 2024-10-22
 
 ### Changed